+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Version 1.10.1
 
+* Hotfix - Login with access token
+* Removed obsolete option, Authenticator
+
+## Version 1.10.0
+
 * Option for open downloading folder
 * New option config - Number download with zero left.
 * Auto ignore attachments downloaded